Safari-611.1.10/LayoutTests/http/tests/workers/navigator-readonly-expected.txt

CONSOLE MESSAGE: Netscape
Test passes if "Netscape" is returned by the worker.
